+++ packages/svelte/src/lib/components/carousel/carousel.anatomy.ts
@@ -1,1 +1,3 @@
-export { anatomy as carouselAnatomy } from '@zag-js/carousel'
+import { anatomy } from '@zag-js/carousel'
+
+export const carouselAnatomy = anatomy.extendWith('progressText')