+++ packages/vue/src/components/carousel/carousel-autoplay-indicator.vue
@@ -1,0 +1,16 @@
+<script lang="ts">
+export type CarouselAutoplayIndicatorBaseProps = {}
+export interface CarouselAutoplayIndicatorProps extends CarouselAutoplayIndicatorBaseProps {}
+</script>
+
+<script setup lang="ts">
+import { useCarouselContext } from './use-carousel-context'
+
+defineProps<CarouselAutoplayIndicatorProps>()
+const carousel = useCarouselContext()
+</script>
+
+<template>
+  <slot v-if="carousel.isPlaying" name="paused" />
+  <slot v-else name="playing" />
+</template>